Tokeknizer refactoring
