+++ lib/parser/deep.js
@@ -7,7 +7,6 @@ const parseContent = (token) => {
     .replace(PARAGRAPH.STRIKE, "<s>$1</s>")
     .replace(PARAGRAPH.UNDERLINE, "<u>$1</u>")
     .replace(PARAGRAPH.LINK, "<a href='$2'>$1</a>")
-
 }
 
 module.exports = parseContent
